# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "NFT Music",
            "url": "https://nft.threenext.com",
            "email": "seki@threenext.com"
        }
    },
    "paths": {
...
        "/collections": {
            "get": {
                "description": "コレクションの情報をリストで取得する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "コレクション"
                ],
                "summary": "コレクションの情報をリストで取得する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ports.CollectionOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "post": {
                "description": "コレクションの情報を1件作成する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "コレクション"
                ],
                "summary": "コレクションの情報を1件作成する",
                "parameters": [
                    {
                        "description": "コレクション",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.CollectionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.CollectionOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
        "/collections/{id}": {
            "get": {
                "description": "コレクションの情報を1件取得する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "コレクション"
                ],
                "summary": "コレクションの情報を1件取得する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "コレクションID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.CollectionOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "put": {
                "description": "コレクションの情報を1件修正する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "コレクション"
                ],
                "summary": "コレクションの情報を1件修正する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "コレクションID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "コレクション",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.CollectionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.CollectionOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "delete": {
                "description": "コレクションの情報を1件削除する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "コレクション"
                ],
                "summary": "コレクションの情報を1件削除する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "コレクションID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ports.CollectionInput": {
            "type": "object",
            "properties": {
                "banner_image_url": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "chain_id": {
                    "type": "integer",
                    "example": 1
                },
                "contract_address": {
                    "type": "string",
                    "example": "0x495f947276749ce646f68ac8c248420045075b34"
                },
                "description": {
                    "type": "string",
                    "example": "私はいつでも明るいです"
                },
                "external_url": {
                    "type": "string",
                    "example": "https://www.yahoo.com"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "建築"
                },
                "royalty": {
                    "type": "integer",
                    "example": 10
                },
                "royalty_receiver": {
                    "type": "string",
                    "example": "0x495f947276749ce646f68ac8c248420045075b34"
                },
                "user_id": {
                    "type": "string",
                    "example": "0193254c-a151-7c4c-b06a-259da7258a27"
                }
            }
        },
        "ports.CollectionOutput": {
            "type": "object",
            "properties": {
                "banner_image_url": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "chain_id": {
                    "type": "integer",
                    "example": 1
                },
                "contract_address": {
                    "type": "string",
                    "example": "0x495f947276749ce646f68ac8c248420045075b34"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-11-04T20:51:26Z"
                },
                "description": {
                    "type": "string",
                    "example": "私はいつでも明るいです"
                },
                "external_url": {
                    "type": "string",
                    "example": "https://www.yahoo.com"
                },
                "id": {
                    "type": "string",
                    "example": "0193254c-a151-7c4c-b06a-259da7258a27"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "建築"
                },
                "royalty": {
                    "type": "integer",
                    "example": 10
                },
                "royalty_receiver": {
                    "type": "string",
                    "example": "0x495f947276749ce646f68ac8c248420045075b34"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-11-04T20:51:26Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "0193254c-a151-7c4c-b06a-259da7258a27"
                }
            }
        },
...
        "ports.ErrorResponseObject": {
            "type": "object",
            "properties": {
                "error_type": {
                    "type": "string",
                    "example": "リクエストの構文が不正やページが見つからない等のタイプ別"
                },
                "message": {
                    "type": "string",
                    "example": "原因となるのエラーメッセージ"
                },
                "status_code": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}